% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telize.R
\name{client_info}
\alias{client_info}
\title{Obtain and geolocate current (external) client IP address}
\usage{
client_info(as_data_frame = TRUE)
}
\arguments{
\item{as_data_frame}{whether to return the results as a data.frame (TRUE)
or a list of named vectors (FALSE). Set to TRUE by default.}
}
\value{
either a data.frame or a list of named vectors, depending on
the option for \code{as_data_frame}
}
\description{
\href{http://www.telize.com}{telize.com} provides IP geolocation based
on MaxMind's free downloadable databases. \code{client_info} uses Telize to determine
the IP address of the current user/client and return the IP address and associated
geolocation info, making it suitable for providing customization (i.e. function
defaults) based on client location data.
}
\note{
This function requires internet access.
}
\examples{
\dontrun{
info <- client_info()
}
}
\author{
Bob Rudis
}

