% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/maxmind.R
\name{maxmind}
\alias{maxmind}
\title{Geolocate IP Addresses through MaxMind Databases}
\usage{
maxmind(ips, file, fields)
}
\arguments{
\item{ips}{a character vector of IP addresses (IPv4 and IPv6 both work)}

\item{file}{the full path to the .mmdb file you want to query.}

\item{fields}{the fields you want to return.}
}
\description{
MaxMind does a set of proprietary geolocation databases
- they're pretty accurate! \code{maxmind} provides a connector to
MaxMind services.
}

