% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rgeolocate.R
\docType{package}
\name{rgeolocate}
\alias{rgeolocate}
\alias{rgeolocate-package}
\title{IP geolocation in R}
\description{
Keep a vest on my chest, I'm sitting in my room as I'm
looking out the face, still got some damage to write about.
}

