% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgeolocate.R
\docType{package}
\name{rgeolocate}
\alias{rgeolocate}
\alias{rgeolocate-package}
\alias{rgeolocate-package}
\title{IP Geolocation in R}
\description{
This package aims to provide connectors to myriad online and offline sources for
taking IP addresses and geolocating them to country, city, timezone and a whole other host of goodies.
For individual connectors, see the package index. It depends on libmaxminddb, which can be obtained and
installed from https://github.com/maxmind/libmaxminddb
}
